Add creating transactions functionality
A user can now create transactions in two different ways, either through the URL 'transactions/new' where they will be asked to select one of their accounts to create
a transaction from, or from the URL 'accounts/:id/transactions/new'. The latter option already contains the account ID in the URL, hence, the user is not asked to
specify an account to transfer from. Details on whether to use account id, sort code, or account number for making transactions have not been decided. Although, they
are trivial changes to make in the code.