Fix currency transfers
There was a bug with currency conversions as money was stored as a base currency of USD which resulted in many floating point errors. The currency and amount transfer logic was completly changed.
Each account now stores the amount in the currency it is in, so less conversions need to be done.