Allow admin to set account ID
To allow the admin to set ID of an account introduces the possibility of ID collisions in the databse.
This was dealt by using server side model validations to prevent this. Now, if an admin inputs a
duplicate ID, they will be warned. If they leave the ID input blank, it will auto-increment
based on the values already in the database. Tests have been added to test this.